docker: don't rebuild the API services
I lost an hour of my life but I've finally understood why the API's
build-arg/env API_COMMIT wasn't always set.

We have 4 API services :

1. the actual API ;
2. the e-mail worker ;
3. the SMS worker ;
4. the scheduler (beat).

They all build the same image so they defined the same build recipe
except for the API which specifies the API_COMMIT.

  api:
    image: notification-api
    build:
      context: api
      dockerfile: api/local/Dockerfile
*     args:
*       API_COMMIT: foobar
  worker:
    image: notification-api
    build:
      context: api
      dockerfile: api/local/Dockerfile
  worker_sms:
    image: notification-api
    build:
      context: api
      dockerfile: api/local/Dockerfile

The problem with this sequence is that they all specify an image and a
build context, which means Docker-Compose will run through each, build
the image and tag it with the image value (here: notification-api).

So my API image was getting built correctly, but then overriden 4 more
times with a new image that didn't have the build arguments, and
therefore when running the API my variable was gone.

The fix is easy: tell the other services to use the API (image:
notification-api) but never build it themselves.